requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2==2.9.6
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
